requirements.txt
torch==2.3.0
torchvision==0.18.0
torchaudio==2.3.0
cython==3.0.10
git+https://github.com/philferriere/cocoapi.git#egg=pycocotools&subdirectory=PythonAPI
git+https://github.com/facebookresearch/fvcore
av==12.0.0
ninja==1.11.1.1
opencv-python==4.9.0.80
pillow==10.3.0
python-dotenv==1.0.1
jellyfish==1.0.3
easyocr==1.7.1
pandas==2.2.2
natsort==8.4.0
nltk==3.8.1
tornado== 6.4